move functions to utils.js and add a sleep function.
