+++ src/main/webapp/simulation_visualization.html
@@ -3,10 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <title>BLETH</title>
-    <script src='simulation_visualization.js'></script>
+    <script type='module' src='simulation_visualization.js'></script>
 </head>
 
 <body onload="visualize()">
-    
+    <table id="real-board-table"></table>
+    <table id="estimated-board-table"></table>
 </body>
 </html>
