+++ src/main/webapp/simulations.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>BLETH</title>
-    <script src='simulations.js'></script>
+    <script type='module' src='simulations.js'></script>
 </head>
 
 <body onload="retrieveSimulations()">
